+++ fastbuilder/i18n/fyne_dedicate.go
@@ -1,0 +1,7 @@
+// +build fyne_gui
+
+package I18n
+
+func Init() {
+	I18nDict = I18nDict_cn
+}
